Fix duplication of timestep events in the simulator.
Previously we were repeating doTimestep with the same array 15 times on
a console line, this results in the array getting stuffed with packet
events and somesuch each time, and it ending up with many, many
duplicates, meaning we have repeated packets and such.

The Emulator probably needs the same treatment, but I'm not familiar
with that code, and it probably deserves a separate commit anyways.